MAINT: collect together check_named_args for discrete & continuous distr
