+++ scipy/stats/tests/test_discrete_basic.py
@@ -1,7 +1,5 @@
 from __future__ import division, print_function, absolute_import
 
-import inspect
-
 import numpy.testing as npt
 import numpy as np
 from scipy.lib.six import xrange
@@ -9,7 +7,8 @@ from scipy.lib.six import xrange
 from scipy import stats
 from common_tests import (check_normalization, check_moment, check_mean_expect,
         check_var_expect, check_skew_expect, check_kurt_expect,
-        check_entropy, check_private_entropy, check_edge_support)
+        check_entropy, check_private_entropy, check_edge_support,
+        check_named_args)
 
 DECIMAL_meanvar = 0  # 1  # was 0
 
@@ -36,7 +35,6 @@ distdiscrete = [
 def test_discrete_basic():
     for distname, arg in distdiscrete:
         distfn = getattr(stats,distname)
-        #npt.assert_(stats.dlaplace.rvs(0.8) is not None)
         np.random.seed(9765456)
         rvs = distfn.rvs(size=2000,*arg)
         supp = np.unique(rvs)
@@ -258,45 +256,6 @@ def check_discrete_chisquare(distfn, arg, rvs, alpha, msg):
 
     npt.assert_(pval > alpha, 'chisquare - test for %s'
            ' at arg = %s with pval = %s' % (msg,str(arg),str(pval)))
-
-
-def check_named_args(distfn, x, shape_args, defaults, meths):
-    """Check calling w/ named arguments."""
-    ### This is a copy-paste from test_continuous_basic.py; dedupe?
-
-    # check consistency of shapes, numargs and _parse signature
-    signature = inspect.getargspec(distfn._parse_args)
-    npt.assert_(signature.varargs is None)
-    npt.assert_(signature.keywords is None)
-    npt.assert_(signature.defaults == defaults)
-
-    shape_argnames = signature.args[1:-len(defaults)]  # self, a, b, loc=0, scale=1
-    if distfn.shapes:
-        shapes_ = distfn.shapes.replace(',', ' ').split()
-    else:
-        shapes_ = ''
-    npt.assert_(len(shapes_) == distfn.numargs)
-    npt.assert_(len(shapes_) == len(shape_argnames))
-
-    # check calling w/ named arguments
-    shape_args = list(shape_args)
-
-    vals = [meth(x, *shape_args) for meth in meths]
-    npt.assert_(np.all(np.isfinite(vals)))
-
-    names, a, k = shape_argnames[:], shape_args[:], {}
-    while names:
-        k.update({names.pop(): a.pop()})
-        v = [meth(x, *a, **k) for meth in meths]
-        npt.assert_array_equal(vals, v)
-        if not 'n' in k.keys():
-            # `n` is first parameter of moment(), so can't be used as named arg
-            npt.assert_equal(distfn.moment(3, *a, **k),
-                             distfn.moment(3, *shape_args))
-
-    # unknown arguments should not go through:
-    k.update({'kaboom': 42})
-    npt.assert_raises(TypeError, distfn.cdf, x, **k)
 
 
 def check_scale_docstring(distfn):
